A bunch of fixes and moving randgen stuff around
