+++ src/NPCRoom.java
@@ -1,3 +1,7 @@
+/**
+ * A room subtype that initializes with an NPC.
+ */
+
 public class NPCRoom extends Room{
     public NPCRoom(GameMap map) {
         super(map);
