+++ src/RandomRoom.java
@@ -2,18 +2,39 @@ import java.util.Random;
 
 /**
  * The RandomRoom class randomly populates a room with an item and an enemy based on chance rolls.
- * Chances: 25% item & enemy, 25% enemy, 50% empty.
+ * Chances: 25% item & enemy, 50% enemy, 25% empty.
  */
 public class RandomRoom extends Room{
     public RandomRoom(GameMap parent){
         super(parent);
         Random rand = new Random();
-        int picker = rand.nextInt(3);
+        int picker = rand.nextInt(4);
         switch(picker) {
             case 0:
                 addItem(makeDrop());
-            case 1:
+            case 1, 2:
                 addEnemy(Enemy.createEnemy());
+        }
+    }
+
+    /**
+     * MakeDrop creates a random item for room generation purposes.
+     * @return The item to add to the room.
+     */
+    public Item makeDrop() {
+        Random rand = new Random();
+        int picker = rand.nextInt(4);
+        switch(picker){
+            case 0:
+                return Armor.createArmor();
+            case 1:
+                return Potion.createPotion();
+            case 2:
+                return Weapon.createWeapon();
+            case 3:
+                return Food.createFood();
+            default:
+                return Potion.createPotion();
         }
     }
 }
